requirements.txt
matplotlib==3.8.2
openai==1.52.0
pandas==2.0.3
python-dotenv==1.0.0
tiktoken==0.3.2
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
accelerate==0.21.0
einops==0.6.1
codegeex==1.0
openpyxl==3.1.2
httpx==0.27.2
fastapi_poe==0.0.56

